Add files via upload
This project is the design and fabrication of a mini elevator system for the partial fulfilment of the course Electronic IOT and Embedded Systems (EIES). The word ‘Mini’ implies that the elevator system would be scaled to a portable size such that it can easily be moved into the electronic laboratory for all necessary test and evaluations, though scaled down dimensionally, it would beyond expectation, effectively give all the functionalities like that of a standards elevator.
Consequently, there a need to fabricate the mechanical parts which includes the carriage, the frame, and the pulley. Then install the Direct current (DC) motor, pullies, and the corresponding brackets. The softwarisation is implemented by the use of a microprocessor (Arduino Uno) and in addition to some other sensors to enable the required feedback and disturbance rejection.
The aim in brevity is to build the plant, implement control through micro-processor and smart sensors and then compare theoretical results with that of the physical results.